+++ .github/CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+🔥 ❗Breaking Changes
 
+- [OF-1806] refac!: Add a new URL property and deprecate others in the AvatarSpaceComponent in https://github.com/magnopus-opensource/connected-spaces-platform/pull/872
+  Simplify the setting and getting of the mesh URL for the avatar component. This is a breaking change due to the removal of the getters and setters for CustomAvatarUrl and AvatarMeshIndex.
 
 ## [6.13.0] - 2025-11-28_20-11-35
 
@@ -24,7 +27,7 @@ All notable changes to this project will be documented in this file.
 
 - [NT-0] feat: Add non-exported iterators to Map by MAG-ElliotMorris in https://github.com/magnopus-opensource/connected-spaces-platform/pull/859
   Add non-exported iterators in support of the SWIG generated interop API work.
-  
+
 ### 🐛 🔨 Bug Fixes
 
 - [OB-4761] fix: unreal crashes on oko asset drag and drop without internet connection by MAG-ChristopherAtkinson in https://github.com/magnopus-opensource/connected-spaces-platform/pull/860